refactor: update app with eslint rules
