+++ src/app/appEntry.tsx
@@ -1,11 +1,11 @@
-import { ThemeProvider } from "@/app/providers/ThemeProvider";
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { Provider } from "react-redux";
-import { store } from "./appStore";
 import { RouterProvider } from "react-router-dom";
-import "@/shared/index.css";
+import { ThemeProvider } from "@/shared/ui/Theme";
 import { appRouter } from "./appRouter";
+import { store } from "./appStore";
+import "@/shared/index.css";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
