+++ src/app/appStore.ts
@@ -1,18 +1,17 @@
-import { newsApi } from "@/entities/news/api/newsApi";
 import { configureStore } from "@reduxjs/toolkit";
-import { useDispatch, TypedUseSelectorHook, useSelector } from "react-redux";
+import { categoryApi } from "@/entities/category";
+import { newsApi } from "@/entities/news";
 import { rootReducer } from "./appReducer";
-import { categoriesApi } from "@/entities/category/api/categoriesApi";
 
 export const store = configureStore({
   reducer: rootReducer,
   middleware: (getDefaultMiddleware) =>
-    getDefaultMiddleware().concat(newsApi.middleware, categoriesApi.middleware),
+    getDefaultMiddleware().concat(
+      newsApi.newsApi.middleware,
+      categoryApi.categoriesApi.middleware
+    ),
 });
 
 export type RootState = ReturnType<typeof store.getState>;
 
 export type AppDispatch = typeof store.dispatch;
-
-export const useAppDispatch: () => AppDispatch = useDispatch;
-export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
